backend/requirement.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic[email]
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
aiosmtplib
python-multipart